// core/types.ts
export interface OfflinePatchSummary {
    name: string;
    version: number;
    nodeCount: number;
    connectionCount: number;
    inputCount: number;
    eventCount: number;
    midiInputCount: number;
    midiOutputCount: number;
}

export interface OfflinePatchValidation {
    ok: true;
    patch: PatchDocument;
    normalizedText: string;
    summary: OfflinePatchSummary;
}